Move misplaced contributors/generators to right namespaces
