+++ src/Castle.Core/DynamicProxy/Contributors/IInvocationCreationContributor.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Castle.DynamicProxy.Generators
+namespace Castle.DynamicProxy.Contributors
 {
 	using System.Reflection;
